Move tokens to graphql_api/token.rs and reverse the order
