--- backend-rust/.sqlx/query-6d712faa16165c7ea829753d3f1562576392ff3d040250e922f1e22b89ac18b7.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n                    SELECT MAX(index) as max_index\n                    FROM tokens\n                ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "max_index",
-        "type_info": "Int8"
-      }
-    ],
-    "parameters": {
-      "Left": []
-    },
-    "nullable": [
-      null
-    ]
-  },
-  "hash": "6d712faa16165c7ea829753d3f1562576392ff3d040250e922f1e22b89ac18b7"
-}
