+++ backend-rust/.sqlx/query-750ba1786f7cec6688c14a0634072b1e5db7c816629a9000b5b243878c4a1323.json
@@ -1,0 +1,20 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "SELECT MAX(index) FROM tokens",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "max",
+        "type_info": "Int8"
+      }
+    ],
+    "parameters": {
+      "Left": []
+    },
+    "nullable": [
+      null
+    ]
+  },
+  "hash": "750ba1786f7cec6688c14a0634072b1e5db7c816629a9000b5b243878c4a1323"
+}
