+++ backend-rust/.sqlx/query-8b160d1140d7823bcaf03711b9934621569d7e66c60aba8cb962054f82a8c2c6.json
@@ -6,7 +6,7 @@
       {
         "ordinal": 0,
         "name": "hash",
-        "type_info": "Varchar"
+        "type_info": "Bpchar"
       },
       {
         "ordinal": 1,
