+++ backend-rust/src/graphql_api/token.rs
@@ -4,11 +4,13 @@ use super::{
 };
 use crate::{
     address::ContractIndex,
+    connection::{ConnectionQuery, DescendingI64},
     scalar_types::{BigInteger, TransactionIndex},
     transaction_event::CisEvent,
 };
-use async_graphql::{ComplexObject, Context, Object, SimpleObject};
+use async_graphql::{connection, ComplexObject, Context, Object, SimpleObject};
 use sqlx::PgPool;
+use tokio_stream::StreamExt;
 
 #[derive(Default)]
 pub struct QueryToken;
@@ -29,6 +31,68 @@ impl QueryToken {
             &token_id,
         )
         .await
+    }
+
+    async fn tokens(
+        &self,
+        ctx: &Context<'_>,
+        #[graphql(desc = "Returns the first _n_ elements from the list.")] first: Option<u64>,
+        #[graphql(desc = "Returns the elements in the list that come after the specified cursor.")]
+        after: Option<String>,
+        #[graphql(desc = "Returns the last _n_ elements from the list.")] last: Option<u64>,
+        #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
+        before: Option<String>,
+    ) -> ApiResult<connection::Connection<String, Token>> {
+        let pool = get_pool(ctx)?;
+        let config = get_config(ctx)?;
+        let query = ConnectionQuery::<DescendingI64>::new(
+            first,
+            after,
+            last,
+            before,
+            config.tokens_connection_limit,
+        )?;
+        let mut row_stream = sqlx::query_as!(
+            Token,
+            "SELECT * FROM (
+                SELECT
+                    index,
+                    init_transaction_index,
+                    total_supply as raw_total_supply,
+                    token_id,
+                    contract_index,
+                    contract_sub_index,
+                    token_address,
+                    metadata_url
+                FROM tokens
+                WHERE tokens.index > $2 AND tokens.index < $1
+                ORDER BY
+                    (CASE WHEN $4 THEN tokens.index END) ASC,
+                    (CASE WHEN NOT $4 THEN tokens.index END) DESC
+                LIMIT $3
+            ) AS token_data
+            ORDER BY token_data.index DESC",
+            query.from.value,
+            query.to.value,
+            query.limit,
+            query.desc
+        )
+        .fetch(pool);
+        let mut connection = connection::Connection::new(false, false);
+        while let Some(token) = row_stream.try_next().await? {
+            connection.edges.push(connection::Edge::new(token.index.to_string(), token));
+        }
+        if let Some(page_max_index) = connection.edges.first() {
+            if let Some(max_index) =
+                sqlx::query_scalar!("SELECT MAX(index) FROM tokens").fetch_one(pool).await?
+            {
+                connection.has_next_page = max_index > page_max_index.node.index;
+            }
+        }
+        if let Some(edge) = connection.edges.last() {
+            connection.has_previous_page = edge.node.index != 0;
+        }
+        Ok(connection)
     }
 }
